More flexible defaults for ssh auth
While the current setup allows users to configure for either
ssh agent use (the default if you don't provide a private key
path), or private keys, it prevents both auth methods from
working in the same project.

Use case: there are a number of developers on a team, some of
whom develop within VMs (and use ssh-agent forwarding so that
they do not need to duplicate credentials), and some of whom
develop without (and directly use the credentails found in
~/.ssh). Now, with no special configuration, both groups of
developers can e.g. push code to GitHub. However, regardless
of the configuration provided for this library, one or the
other group would be unable to deploy.

With this change, such a team could leave out any credentails
confirguration and both setups would work correctly.

For completeness, I have also provided an explicit option for
specifying an ssh agent.